requirements.txt
google-generativeai
selenium
webdriver-manager